Python_code/requirements.txt
serial
bleak
PyQt5
numpy
pyqtgraph
PyOpenGL